+++ cli/Sources/TuistServer/Client/Client+Server.swift
@@ -1,6 +1,5 @@
 import Foundation
 import OpenAPIRuntime
-import OpenAPIURLSession
 import TuistHTTP
 
 extension Client {
@@ -10,8 +9,8 @@ extension Client {
     public static func authenticated(serverURL: URL, authenticationURL: URL? = nil) -> Client {
         .init(
             serverURL: serverURL,
-            transport: URLSessionTransport(configuration: .init(session: .tuistShared)),
-            middlewares: [
+            transport: TuistURLSessionTransport(),
+            middlewares: HARRecordingMiddlewareFactory.middlewares() + [
                 RequestIdMiddleware(),
                 ServerClientCLIMetadataHeadersMiddleware(),
                 ServerClientAuthenticationMiddleware(authenticationURL: authenticationURL),
@@ -25,8 +24,8 @@ extension Client {
     public static func unauthenticated(serverURL: URL) -> Client {
         .init(
             serverURL: serverURL,
-            transport: URLSessionTransport(configuration: .init(session: .tuistShared)),
-            middlewares: [
+            transport: TuistURLSessionTransport(),
+            middlewares: HARRecordingMiddlewareFactory.middlewares() + [
                 RequestIdMiddleware(),
                 VerboseLoggingMiddleware(),
                 OutputWarningsMiddleware(),